+++ gha_utils/metadata.py
@@ -206,73 +206,73 @@ nuitka_matrix={
             "os": "ubuntu-24.04-arm",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-linux-arm64-346ce66.bin"
+            "bin_name": "mpm-linux-arm64.bin"
         },
         {
             "os": "ubuntu-24.04-arm",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-linux-arm64-6f27db4.bin"
+            "bin_name": "mpm-linux-arm64.bin"
         },
         {
             "os": "ubuntu-24.04",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-linux-x64-346ce66.bin"
+            "bin_name": "mpm-linux-x64.bin"
         },
         {
             "os": "ubuntu-24.04",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-linux-x64-6f27db4.bin"
+            "bin_name": "mpm-linux-x64.bin"
         },
         {
             "os": "macos-26",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-macos-arm64-346ce66.bin"
+            "bin_name": "mpm-macos-arm64.bin"
         },
         {
             "os": "macos-26",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-macos-arm64-6f27db4.bin"
+            "bin_name": "mpm-macos-arm64.bin"
         },
         {
             "os": "macos-15-intel",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-macos-x64-346ce66.bin"
+            "bin_name": "mpm-macos-x64.bin"
         },
         {
             "os": "macos-15-intel",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-macos-x64-6f27db4.bin"
+            "bin_name": "mpm-macos-x64.bin"
         },
         {
             "os": "windows-11-arm",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-windows-arm64-346ce66.bin"
+            "bin_name": "mpm-windows-arm64.bin"
         },
         {
             "os": "windows-11-arm",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-windows-arm64-6f27db4.bin"
+            "bin_name": "mpm-windows-arm64.bin"
         },
         {
             "os": "windows-2025",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
-            "bin_name": "mpm-windows-x64-346ce66.exe"
+            "bin_name": "mpm-windows-x64.exe"
         },
         {
             "os": "windows-2025",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
-            "bin_name": "mpm-windows-x64-6f27db4.exe"
+            "bin_name": "mpm-windows-x64.exe"
         },
         {"state": "stable"}
     ]
@@ -1983,7 +1983,7 @@ class Metadata:
             # so we need a copy the main variants it corresponds to.
             bin_name_include = {k: variations[k] for k in matrix.variations}
             bin_name_include["bin_name"] = (
-                "{cli_id}-{target}-{short_sha}.{extension}"
+                "{cli_id}-{target}.{extension}"
             ).format(**variations)
             matrix.add_includes(bin_name_include)